+++ client/client_state.C
@@ -551,12 +551,12 @@ PROJECT* CLIENT_STATE::lookup_project(const char* master_url) {
     int len1, len2;
     char *mu;
 
-    len1 = strlen(master_url);
+    len1 = (int)strlen(master_url);
     if (master_url[strlen(master_url)-1] == '/') len1--;
 
     for (unsigned int i=0; i<projects.size(); i++) {
         mu = projects[i]->master_url;
-        len2 = strlen(mu);
+        len2 = (int)strlen(mu);
         if (mu[strlen(mu)-1] == '/') len2--;
         if (!strncmp(master_url, projects[i]->master_url, max(len1,len2))) {
             return projects[i];
