+++ client/client_types.C
@@ -875,7 +875,7 @@ const char* FILE_INFO::get_init_url(bool is_upload) {
     start_url = current_url;
     while(1) {
         if (!is_correct_url_type(is_upload, urls[current_url])) {
-            current_url = (current_url + 1)%urls.size();
+            current_url = (current_url + 1)%((int)urls.size());
             if (current_url == start_url) {
                 msg_printf(project, MSG_ERROR,
                     "Couldn't find suitable URL for %s", name);
@@ -894,7 +894,7 @@ const char* FILE_INFO::get_init_url(bool is_upload) {
 const char* FILE_INFO::get_next_url(bool is_upload) {
     if (!urls.size()) return NULL;
     while(1) {
-        current_url = (current_url + 1)%urls.size();
+        current_url = (current_url + 1)%((int)urls.size());
         if (current_url == start_url) {
             return NULL;
         }
@@ -1002,7 +1002,7 @@ int FILE_INFO::gzip() {
     if (!in) return ERR_FOPEN;
     gzFile out = gzopen(outpath, "wb");
     while (1) {
-        int n = fread(buf, 1, BUFSIZE, in);
+        int n = (int)fread(buf, 1, BUFSIZE, in);
         if (n <= 0) break;
         int m = gzwrite(out, buf, n);
         if (m != n) {
@@ -1479,7 +1479,7 @@ int RESULT::write(MIOFILE& out, bool to_server) {
             wup->version_num
         );
     }
-    n = stderr_out.length();
+    n = (int)stderr_out.length();
     if (n || to_server) {
         out.printf("<stderr_out>\n");
